@@ -7,41 +7,44 @@
     "@storybook/addon-knobs": "^5.0.11",
     "@storybook/addon-options": "^5.0.11",
     "@storybook/react": "^5.0.11",
-    "@types/jest": "23.3.9",
-    "@types/node": "10.12.7",
-    "@types/react": "16.7.6",
-    "@types/react-dom": "16.0.9",
-    "@types/react-intl": "^2.3.14",
-    "@types/react-redux": "^6.0.9",
-    "@types/react-router": "^4.4.1",
-    "@types/react-router-dom": "^4.3.1",
-    "@types/react-router-redux": "^5.0.16",
+    "@types/jest": "^24.0.12",
+    "@types/lodash": "^4.14.123",
+    "@types/node": "^11.13.8",
+    "@types/react": "^16.8.15",
+    "@types/react-dom": "^16.8.4",
+    "@types/react-intl": "^2.3.17",
+    "@types/react-redux": "^7.0.1",
+    "@types/react-router": "^4.4.5",
+    "@types/react-router-dom": "^4.3.2",
+    "@types/react-router-redux": "^5.0.18",
     "aphrodite": "^2.3.1",
     "awesome-typescript-loader": "^5.2.1",
+    "chokidar-cli": "^1.2.2",
     "cross-env": "^5.2.0",
-    "hypernova": "^2.4.0",
+    "hypernova": "^2.5.0",
     "hypernova-amp": "^3.2.0",
     "hypernova-react": "^2.1.0",
     "isomorphic-fetch": "^2.2.1",
-    "lodash-es": "^4.17.11",
+    "lodash": "^4.17.11",
     "node-noop": "^1.0.0",
     "nodemon": "^1.19.0",
-    "raw-loader": "^0.5.1",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-helmet": "^5.2.0",
-    "react-intl": "^2.7.2",
-    "react-redux": "^5.1.1",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
+    "raw-loader": "^2.0.0",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "react-helmet": "^5.2.1",
+    "react-intl": "^2.8.0",
+    "react-redux": "^7.0.3",
+    "react-router": "^5.0.0",
+    "react-router-dom": "^5.0.0",
     "react-scripts": "^3.0.0",
-    "react-toastify": "^4.5.1",
+    "react-toastify": "^5.1.0",
     "redux": "^4.0.1",
-    "redux-devtools-extension": "^2.13.5",
+    "redux-devtools-extension": "^2.13.8",
     "typescript": "^3.4.5",
+    "uglifyjs-webpack-plugin": "^2.1.2",
     "url-join": "^4.0.0",
-    "webpack": "4.19.1",
-    "webpack-bundle-analyzer": "^3.0.3"
+    "webpack": "^4.30.0",
+    "webpack-bundle-analyzer": "^3.3.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -68,19 +71,18 @@
     "not op_mini all"
   ],
   "devDependencies": {
-    "@types/lodash-es": "^4.17.1",
-    "@types/react-helmet": "^5.0.7",
+    "@types/react-helmet": "^5.0.8",
     "@types/storybook__addon-actions": "^3.4.2",
     "@types/storybook__addon-knobs": "^5.0.0",
     "@types/storybook__addon-options": "^4.0.1",
     "@types/storybook__react": "^4.0.1",
-    "node-sass": "^4.11.0",
+    "node-sass": "^4.12.0",
     "plop": "^2.3.0",
     "sass-loader": "^7.1.0",
     "tslint": "^5.16.0",
     "tslint-config-prettier": "^1.18.0",
     "tslint-react": "^4.0.0",
-    "webpack-cli": "^3.1.2"
+    "webpack-cli": "^3.3.1"
   },
   "rescripts": [
     "env"
